Moved the cohort functions to their own file.
